% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProps.R
\name{calcProps}
\alias{calcProps}
\title{emCalcMix(x,densities,p,maxiter,conv)}
\usage{
calcProps(data)
}
\arguments{
\item{data}{A data frame with a 'class' column, containing class values
for each row. Defaults to NULL.}
}
\value{
classProps A vector of class proportions for the data$class column.
}
\description{
This function calculates teh class proportions of the 'class' column
for the data set.  Data set must contain a column named class.
}
\examples{
##data is a data frame containing a column named 'class'.
p <- classProps(data)

}
\author{
Jennifer Starling
}
\keyword{}

