% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanData}
\alias{cleanData}
\title{cleanData(train,test)}
\usage{
cleanData(train, test)
}
\arguments{
\item{train}{A data set whose class values are known. Defaults to NULL.}

\item{test}{A data set whose class values are unknown.  Defaults to NULL.}
}
\value{
A list containing the following components:

cleanData[[1]] = Cleaned training data set.

cleanData[[2]] = Cleaned test training set.
}
\description{
This function cleans a test and a training data set, to eliminate features
which cannot be used for classification due to problematic or unclean data.
Eliminates columns (features) where stdev=NA or 0.  Eliminates rows with
infinite values.
}
\examples{
clean_data <- cleanData(train=dataset1, test=dataset2)

}
\author{
Jennifer Starling
}
\keyword{}

