% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation_for_roxygen2.R
\docType{data}
\name{em_by_feature}
\alias{em_by_feature}
\title{List containing output of the EM algorithm run for each individual feature.}
\format{A list object with four elements:
\describe{
  \item{$est_mixing_props}{Matrix of estimated class proportions (cols) for each feature (rows)}
  \item{$feature_MSEs}{Vector of MSEs for each feature, for estimated props. vs unknown props.}
  \item{$feature_iter}{Vector of number of iterations to EM convergence for each feature.}
  \item{$feature_log}{List of vectors of log-likelihood functions for each feature..}
  ...
}}
\usage{
em_by_feature
}
\description{
This object is the result of running emAllFeatures function for data sets each containing
the class column and the single feature.  Generated using maxiter=1000, conv=.00001.
}
\keyword{datasets}

