% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emAllFeatures.R
\name{emAllFeatures}
\alias{emAllFeatures}
\title{emAllFeatures(known,unknown,p=rep(.2,5),maxiter=2000,conv=.00001)}
\usage{
emAllFeatures(known, unknown, p = rep(0.2, 5), maxiter = 2000,
  conv = 1e-05)
}
\arguments{
\item{known}{A data set with class as the first column.  This data set has known 
class proportions.}

\item{unknown}{A data set whose proportions will be estimated.  This data set may have 
class as the first column, to return an MSE to compare the estimates to the actual proportions.}

\item{p}{Vector of class starting proportion estimates.  Default is .2 for 5 classes.}

\item{maxiter}{Maximum number of iterations of the EM algorithm allowed. Default is 2000.}

\item{conv}{Convergence criteria for change in likelihood function.  Change between}
}
\value{
A list containing the following components:

p = vector of estimated mixing proportions for the unknown data set.

iter = number of iterations before EM algorithm converged or was stopped.

allp = (iter x 5) data frame containing all phat iterated estimates.

logl = Vector of log-likelihood values, with one log-lik value for each iteration.

mse = MSE of actual unknown class proportions (if in data set) versus estimated.
}
\description{
This function performs the Expectation-Maximization Algorithm to estimate
a vector of class proportions for a single feature in a data set wih an 
unknown set of class proportions. Called by \code{\link{featureMixtureProportion}}.
}
\examples{

## Define ctrl object.
em <- emAllFeatures(df1,df2,p=rep(.2,5),maxiter=1000,conv=.00001)

}
\author{
Jennifer Starling
}
\keyword{EM}
\keyword{expectation-maximization}

