% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emCalcMix.R
\name{emCalcMix}
\alias{emCalcMix}
\title{emCalcMix(x,densities,p,maxiter,conv)}
\usage{
emCalcMix(x, densities, p = rep(0.2, 5), maxiter = 2000, conv = 1e-05)
}
\arguments{
\item{x}{A vector of observed feature values for specified feature. Defaults to NULL.}

\item{densities}{List of k density function; one per class. Flexible length, which must 
equal k.}

\item{p}{Vector of class starting proportion estimates.  Default is .2 for 5 classes.}

\item{maxiter}{Maximum number of iterations of the EM algorithm allowed. Default is 2000.}

\item{conv}{Convergence criteria for change in likelihood function.  Change between}
}
\value{
A list containing the following components:

p = vector of 5 estimated mixing proportions for feature.

iter = number of iterations before algorithm converged or was stopped.

allp = (iter x 5) data frame containing all phat iterated estimates.

logl = Vector of log-likelihood values, with one log-lik value for each iteration.
}
\description{
This function performs the Expectation-Maximization Algorithm to estimate
a vector of class proportions for a single feature in a data set wih an 
unknown set of class proportions. Called by \code{\link{featureMixtureProportion}}.
}
\examples{

## Define ctrl object.
em <- emCalcMix(f_sample,densities,p=rep(.2,5),maxiter=1000,conv=.00001)

##Sample density function:
f_class1 = df[df$class=="eb",f_name]
    	kd1  = density(f_class1)
    	dens1 = function(x)
	{
		min_den <- quantile(kd1$y,.05)
		de <- approx(kd1$x,kd1$y,x)$y
		de[is.na(de)] <- min_den
		return(de)
	}

}
\author{
Jennifer Starling
}
\keyword{EM}
\keyword{expectation-maximization}

