% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation_for_roxygen2.R
\docType{data}
\name{est_mixing_props_hip}
\alias{est_mixing_props_hip}
\title{List of k estimated class proportions for each feature of the hip data set.}
\format{A list object with 2118 elements; each element is vector of length k:
\describe{
  \item{list[[1]]}{Vector of estimated class proportions for feature 1}
  \item{list[[2]]}{Vector of estimated class proportions for feature 2}
  ...
}}
\usage{
est_mixing_props_hip
}
\description{
This object is the result of running the featureMixtureProportion(hip)
function.  It is designed to streamline the code run time for analysis of hip,
as the function takes some time to compute.
}
\keyword{datasets}

