% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation_for_roxygen2.R
\docType{data}
\name{est_mixing_props_linear}
\alias{est_mixing_props_linear}
\title{List of k estimated class proportions for each feature of the linear data set.}
\format{A list object with 5509 elements; each element is vector of length k:
\describe{
  \item{list[[1]]}{Vector of estimated class proportions for feature 1}
  \item{list[[2]]}{Vector of estimated class proportions for feature 2}
  ...
}}
\usage{
est_mixing_props_linear
}
\description{
This object is the result of running the featureMixtureProportion(linear)
function.  It is designed to streamline the code run time for analysis of
linear, as the function takes some time to compute.
}
\keyword{datasets}

