% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation_for_roxygen2.R
\docType{data}
\name{hip}
\alias{hip}
\title{Hipparcos variable star data.}
\format{A data frame with 2118 rows and 144 variables:
\describe{
  \item{class}{variable star classification}
  ...
}}
\source{
\url{http://adsabs.harvard.edu/abs/2012ada..confE..15S}
}
\usage{
hip
}
\description{
A dataset containing the classes and other features of 2118
variable stars.
}
\keyword{datasets}

