% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation_for_roxygen2.R
\docType{data}
\name{linear}
\alias{linear}
\title{LINEAR variable star data.}
\format{A data frame with 5509 rows and 144 variables:
\describe{
  \item{class}{variable star classification}
  ...
}}
\source{
\url{http://adsabs.harvard.edu/abs/2012ada..confE..15S}

\url{http://iopscience.iop.org/1538-3881/146/4/101/suppdata/aj482492t5_mrt.txt}
}
\usage{
linear
}
\description{
A dataset containing the classes and other features of 2118
variable stars.  The features are derived from the periodic light
curves, with raw data available in the second source link.
}
\keyword{datasets}

