% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFSampleScheme.R
\name{RFSampleScheme}
\alias{RFSampleScheme}
\title{RFSampleScheme: A Random Forest function.}
\usage{
RFSampleScheme(.totalObs, .minObsPerClass, .nKnown, .propClassKnown, .nUnknown,
  .propClassUnknown)
}
\arguments{
\item{totalObs}{The number of observatiosn desired for the entire data set, all classes together.
Defaults to NULL.}

\item{minObsPerClass}{The minimum number of observations allowed per class.  Defaults to NULL.}

\item{nKnown}{Number of observataions in the data set with known class labels. Defaults to NULL.}

\item{nUnknown}{Number of observataions in the data set with unknown class labels. Defaults to NULL.}

\item{propClassUnknown}{Estimated class proportiosnin the data set with unknown class labels.
Defaults to NULL.}
}
\value{
A vector containing sampsize class proportions to use in the RandomForest model.
}
\description{
This function builds a vector of sample sizes to be used as the sampsize parameter
in the random forest function.  The vector contains sample sizes for each class level.
}
\author{
Robert Saw
}
\keyword{classification}
\keyword{forest}
\keyword{random}
\keyword{star}
\keyword{variable}

