% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation_for_roxygen2.R
\docType{data}
\name{sample_model}
\alias{sample_model}
\title{Sample model using weighted data, out of the box random forest, EM feature selection.
Error rate of 18% for classifying linear data using hip data.}
\format{A list object containing model components.}
\usage{
sample_model
}
\description{
Sample model using weighted data, out of the box random forest, EM feature selection.
Error rate of 18% for classifying linear data using hip data.
}
\keyword{datasets}

