% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeClassSim.R
\name{threeClassSim}
\alias{threeClassSim}
\title{threeClassSim(nA,nB,nC)}
\usage{
threeClassSim(nA, nB, nC, wellSep = TRUE)
}
\arguments{
\item{nA}{Number of observations to generate from class A. Defaults to NULL.}

\item{nB}{Number of observations to generate from class B. Defaults to NULL.}

\item{nC}{Number of observations to generate from class C. Defaults to NULL.}

\item{wellSep}{Value to indicate whether distributions of three classes should be well separated.  Defaults to TRUE.}
}
\value{
A list containing the following components:

x$data The simulated data set.

x$key The distribution key for the simulated data set.
}
\description{
Creates a simulated data set with three classes (A,B,C).
}
\examples{

## Define ctrl object.
datasim <- threeClassSim(100,200,300)


}
\author{
Jennifer Starling
}

