% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeClassSimCor.R
\name{threeClassSimCor}
\alias{threeClassSimCor}
\title{threeClassSimCor(nA,nB,nC)}
\usage{
threeClassSimCor(nA, nB, nC, corFeatures, wellSep = TRUE)
}
\arguments{
\item{nA}{Number of observations to generate from class A. Defaults to NULL.}

\item{nB}{Number of observations to generate from class B. Defaults to NULL.}

\item{nC}{Number of observations to generate from class C. Defaults to NULL.}

\item{wellSep}{Value to indicate whether distributions of three classes should be well separated.  Defaults to TRUE.}
}
\value{
A list containing the following components:

x$data The simulated data set.

x$key The distribution key for the simulated data set.
}
\description{
Creates a simulated data set with three classes (A,B,C).
Some features are corrupted, to simulate real-world data.
}
\examples{

## Define ctrl object.
datasim <- threeClassSimCor(100,200,300)


}
\author{
Jennifer Starling
}

