\name{varstar-package}
\alias{varstar-package}
\alias{varstar}
\docType{package}
\title{
\packageTitle{varstar}
}
\description{
\packageDescription{varstar}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{varstar}
\packageIndices{varstar}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{varstar}

Maintainer: \packageMaintainer{varstar}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
