% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureMixtureProportion.R
\name{featureMixtureProportion}
\alias{featureMixtureProportion}
\title{featureMixtureProportion(df)}
\usage{
featureMixtureProportion(df)
}
\arguments{
\item{df}{A data frame containing rows of observations.  The first column must be
class.  Other columns contain features. Defaults to NULL.}
}
\value{
A list containing one element for each feature in the data set.

Each list element is a vector of k mixing proportions.
}
\description{
This function calls the \code{\link{emCalcMix}} function repeatedly, for each 
feature (column) in the data set given. Creates estimated class proportion vector 
for each feature. Data set must have 'class' as the first column.
}
\examples{
est_mixing_proportions <- featureMixtureProportion(df=unknown)

}
\author{
Jennifer Starling
}
\keyword{EM}
\keyword{expectation-maximization}

